+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Fixed bug in ilo_https_cert resource: Accept `ILO_SDK::Client` object for ilo property.
  - Fixed bug in ilo_firmware_update resource: Accept String for fw_version property.
  - Fix other various property restrictions,
+ - Fixed bug in ilo_https_cert :import when importing from file.
 
 ### 1.3.0
  - Refactored ilo_bios resource to support all settings. (Breaking Change)